+++ source/js/dialogs.js
@@ -12,9 +12,9 @@ import escapeHtml from 'escape-html';
 import {spinner} from './spinner.js';
 
 /**
- *  The version of FHIR output by this app (in the "show" dialogs).
+ *  The default version of FHIR output by this app (in the "show" dialogs).
  */
-const outputFHIRVersion_ = 'R4';
+const defaultVersion = 'R4';
 
 /**
  *  A reference to the element into which the form will be placed.
@@ -88,7 +88,7 @@ export const Dialogs = {
    */
   showFHIRSDCQuestionnaire: function (event) {
     var sdc = LForms.Util.getFormFHIRData('Questionnaire',
-      outputFHIRVersion_, formContainer_);
+      fhirService.getFHIRVersion() || this.defaultVersion, formContainer_);
     var fhirString = JSON.stringify(sdc, null, 2);
     this.showDataDialog("FHIR SDC Questionnaire Content", fhirString);
   },
@@ -100,7 +100,7 @@ export const Dialogs = {
    */
   showFHIRSDCQuestionnaireResponse: function (event) {
     var sdc = LForms.Util.getFormFHIRData('QuestionnaireResponse',
-      outputFHIRVersion_, formContainer_, {
+      fhirService.getFHIRVersion() || this.defaultVersion, formContainer_, {
         subject: fhirService.getCurrentPatient()
       });
     if (fhirService.getCurrentUser() && !sdc.author) {